% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.compCohensD}
\alias{.compCohensD}
\title{Compute Cohen's d}
\usage{
.compCohensD(t, n)
}
\arguments{
\item{t}{t-value}

\item{n}{sample size per group}
}
\description{
Compute Cohen's d from t-value with equal sized groups of size n
}
