% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.compCohensDData}
\alias{.compCohensDData}
\title{Compute Cohen's d from the observed data}
\usage{
.compCohensDData(control, treatment)
}
\arguments{
\item{control}{values of the control group}

\item{treatment}{values of the treatment group}
}
\description{
Compute Cohen's d from the observed data
}
