% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.compCohensDSE}
\alias{.compCohensDSE}
\title{Compute the standard error of Cohen's d}
\usage{
.compCohensDSE(d, n1, n2)
}
\arguments{
\item{d}{Cohen's d}

\item{n1}{Sample size in group 1}

\item{n2}{Sample size in group 2}
}
\description{
Compute the standard error of Cohen's d
}
