% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.compCohensDStat}
\alias{.compCohensDStat}
\title{Compute Cohen's d from a test statistic and group sizes}
\usage{
.compCohensDStat(statistic, n1, n2)
}
\arguments{
\item{statistic}{Test statistic}

\item{n1}{Sample size in group 1}

\item{n2}{Sample size in group 2}
}
\description{
Compute Cohen's d from a test statistic and group sizes
}
