% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.compR2t}
\alias{.compR2t}
\title{Compute R squared for the t-test}
\usage{
.compR2t(x, y)
}
\arguments{
\item{x}{values of group 1}

\item{y}{values of group 2}
}
\description{
Compute R squared for the t-test
}
