% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositeScores.R
\name{.compscoreHack}
\alias{.compscoreHack}
\title{P-Hacking function for scale redefinition / Composite Scores}
\usage{
.compscoreHack(df, dv, compv, ndelete, strategy = "firstsig", alpha = 0.05)
}
\arguments{
\item{df}{Data frame containing dependent variable and composite score items as columns}

\item{dv}{Location of dependent variable in the data frame}

\item{compv}{Location of composite score variables in the data frame}

\item{ndelete}{How many items should be deleted from the scale at maximum?}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{alpha}{Significance level of the t-test (default: 0.05)}
}
\description{
P-Hacking function for scale redefinition / Composite Scores
}
