% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploitCutoffs.R
\name{.cutoffHack}
\alias{.cutoffHack}
\title{P-Hacking function for exploiting cutoff values}
\usage{
.cutoffHack(df, iv, dv, ambitious = FALSE, alpha = 0.05)
}
\arguments{
\item{df}{Data frame with one continuous independent variable and one continuous dependent variable}

\item{iv}{Location of the independent variable in the data frame}

\item{dv}{Location of the dependent variable in the data frame}

\item{ambitious}{Ambitious p-hacking (smallest p value): TRUE/FALSE}

\item{alpha}{Significance level of the t-test}
}
\description{
P-Hacking function for exploiting cutoff values
}
