% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.draw.study.effect}
\alias{.draw.study.effect}
\title{Draw one study-level effect size}
\usage{
.draw.study.effect(effect = 0, heterogeneity = 0)
}
\arguments{
\item{effect}{Mean effect size across studies}

\item{heterogeneity}{Between-study heterogeneity}
}
\description{
Draw one study-level effect size
}
