% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favorableImputation.R
\name{.easyimpute}
\alias{.easyimpute}
\title{Single value imputation function
Imputes NA values in a single variable using the function specified in fun}
\usage{
.easyimpute(x, fun, ...)
}
\arguments{
\item{x}{The target variable of the imputation}

\item{fun}{The function used to replace missing values that takes x as an argument (e.g., mean)}

\item{...}{Additional arguments to fun}
}
\description{
Single value imputation function
Imputes NA values in a single variable using the function specified in fun
}
