% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favorableImputation.R
\name{.estimate_mode}
\alias{.estimate_mode}
\title{Estimate mode of continuous variables
Estimates mode of continuous variables using the density() function}
\usage{
.estimate_mode(x)
}
\arguments{
\item{x}{The target variable for which the mode should be searched}
}
\description{
Estimate mode of continuous variables
Estimates mode of continuous variables using the density() function
}
