% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.extractoutlier}
\alias{.extractoutlier}
\title{Create data frames without outliers}
\usage{
.extractoutlier(x, y, outsx, outsy)
}
\arguments{
\item{x}{Original vector of x values}

\item{y}{Original vector of y values}

\item{outsx}{Outlier values to be removed from x}

\item{outsy}{Outlier values to be removed from y}
}
\description{
Inputs data frame and two sets of outlier values, outputs list with three data frames
}
