% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.fisherz_to_r}
\alias{.fisherz_to_r}
\title{Convert a Fisher-z value to a correlation}
\usage{
.fisherz_to_r(effect)
}
\arguments{
\item{effect}{Fisher-z effect size}
}
\description{
Convert a Fisher-z value to a correlation
}
