% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favorableImputation.R
\name{.impHack}
\alias{.impHack}
\title{P-Hacking function favorable imputation in univariate linear regression}
\usage{
.impHack(df, x, y, which = c(1:10), strategy = "firstsig",
  alpha = 0.05, seed = 1234)
}
\arguments{
\item{df}{Data frame containing x and y variables as columns}

\item{x}{Location of x variable (predictor) in the data frame}

\item{y}{Location of y variable (criterion) in the data frame}

\item{which}{Which missing value handling method? 1: delete missing, 2: mean imputation, 3: median imputation, 4: mode imputation, 5: predictive mean matching, 6: weighted predictive mean matching, 7: sample from observed values, 8: Bayesian linear regression, 9: linear regression ignoring model error, 10: linear regression predicted values}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{alpha}{Significance level of the t-test (default: 0.05)}

\item{seed}{Random seed for imputation processes that require a random process}
}
\description{
Outputs a p-hacked p-value and a vector of all p-values that were computed in the process
}
