% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favorableImputation.R
\name{.miceNoOutput}
\alias{.miceNoOutput}
\title{Use mice::mice function without output being sent to the console}
\usage{
.miceNoOutput(...)
}
\arguments{
\item{...}{All arguments sent to the mice function}
}
\description{
Use mice::mice function without output being sent to the console
}
