% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectiveReportingIV.R
\name{.multIVhack}
\alias{.multIVhack}
\title{P-Hacking function for multiple independent variables}
\usage{
.multIVhack(df, ivs, control, ambitious = FALSE,
  alternative = "two.sided")
}
\arguments{
\item{df}{Data frame with one group variable and multiple dependent variables}

\item{ivs}{Vector defining the IV columns (will be checked in given order)}

\item{control}{Scalar defining location of control group variable}

\item{ambitious}{Ambitious p-hacking (smallest p value): TRUE/FALSE}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater")}
}
\description{
Outputs a p-hacked p-value and a vector of all p-values that were computed in the process
}
