% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectiveReportingIV.R
\name{.multIVhack_ttest}
\alias{.multIVhack_ttest}
\title{P-Hacking function for multiple independent variables in a t-test}
\usage{
.multIVhack_ttest(
  df,
  ivs,
  control,
  strategy = "firstsig",
  alternative = "two.sided",
  alpha = 0.05
)
}
\arguments{
\item{df}{Data frame (wide format) containing a control group variable and multiple treatment group variables}

\item{ivs}{Location of the independent variables (treatment groups) in the (wide) data frame}

\item{control}{Location of the control group in the (wide) data frame}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater"). For the t-test path, \code{"greater"} tests whether the treatment or second group exceeds the control or first group.}

\item{alpha}{Significance level of the t-test (default: 0.05)}
}
\description{
Returns a p-hacked p-value and a vector of all p-values that were computed in the process
}
