% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.onesided_from_twosided}
\alias{.onesided_from_twosided}
\title{Convert a two-sided p-value to a one-sided p-value}
\usage{
.onesided_from_twosided(p.twosided, stat, alternative = "two.sided")
}
\arguments{
\item{p.twosided}{Two-sided p-value}

\item{stat}{Test statistic}

\item{alternative}{Direction of the test}
}
\description{
Convert a two-sided p-value to a one-sided p-value
}
