% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierExclusion.R
\name{.out.boxplot}
\alias{.out.boxplot}
\title{Box plot outlier definition}
\usage{
.out.boxplot(x, y)
}
\arguments{
\item{x}{Vector of values from which outliers should be excluded}

\item{y}{Vector of values from which outliers should be excluded}
}
\description{
Box plot outlier definition function
}
