% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierExclusion.R
\name{.out.cook}
\alias{.out.cook}
\title{Cook's Distance outlier definition}
\usage{
.out.cook(x, y)
}
\arguments{
\item{x}{Vector of x values (predictor in linear regression)}

\item{y}{Vector of y values (criterion in linear regression)}
}
\description{
Excludes values that have a Cook's distance larger than the median of an F distribution with p and n-p degrees of freedom or larger than 1 (see Wikipedia for Cook's distance for the cutoff)
}
