% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierExclusion.R
\name{.out.covratio}
\alias{.out.covratio}
\title{Covariance ratio outlier definition}
\usage{
.out.covratio(x, y)
}
\arguments{
\item{x}{Vector of x values (predictor in linear regression)}

\item{y}{Vector of y values (criterion in linear regression)}
}
\description{
Excludes values that have a covariance ratio differing from 1 (cutoff: influence.measues function internal)
}
