% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierExclusion.R
\name{.out.dfbeta}
\alias{.out.dfbeta}
\title{DFBETAS outlier definition}
\usage{
.out.dfbeta(x, y)
}
\arguments{
\item{x}{Vector of x values (predictor in linear regression)}

\item{y}{Vector of y values (criterion in linear regression)}
}
\description{
Excludes the 1-3 values that have the highest influence on the regression slope
}
