% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierExclusion.R
\name{.out.dffits}
\alias{.out.dffits}
\title{DFFITS outlier definition}
\usage{
.out.dffits(x, y)
}
\arguments{
\item{x}{Vector of x values (predictor in linear regression)}

\item{y}{Vector of y values (criterion in linear regression)}
}
\description{
Excludes values that have absolute DFFIT values larger than 2*sqrt(2/n) (see Wikipedia page for DFFITS for the cutoff)
}
