% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierExclusion.R
\name{.out.leverage}
\alias{.out.leverage}
\title{Leverage values outlier definition}
\usage{
.out.leverage(x, y)
}
\arguments{
\item{x}{Vector of x values (predictor in linear regression)}

\item{y}{Vector of y values (criterion in linear regression)}
}
\description{
Excludes values that have high leverage values (3 times larger than the mean leverage value 3*(p/n), see 'https://newonlinecourses.science.psu.edu/stat501/node/338/' for the cutoff)
}
