% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierExclusion.R
\name{.out.mahalanobis}
\alias{.out.mahalanobis}
\title{Robust Mahalanobis Distance outlier definition}
\usage{
.out.mahalanobis(x, y)
}
\arguments{
\item{x}{Vector of x values (predictor in linear regression)}

\item{y}{Vector of y values (criterion in linear regression)}
}
\description{
Excludes values that have a high robust Mahalanobis Distance (cutoff: squared MD > qchisq(0.98, 2), see Filzmoser et al. (2005))
}
