% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierExclusion.R
\name{.out.percentrule}
\alias{.out.percentrule}
\title{Percentage outlier definition}
\usage{
.out.percentrule(x, y)
}
\arguments{
\item{x}{Vector of values from which outliers should be excluded}

\item{y}{Vector of values from which outliers should be excluded}
}
\description{
Percentage outlier definition function
}
