% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierExclusion.R
\name{.out.residual}
\alias{.out.residual}
\title{Residuals outlier definition}
\usage{
.out.residual(x, y, type)
}
\arguments{
\item{x}{Vector of x values (predictor in linear regression)}

\item{y}{Vector of y values (criterion in linear regression)}

\item{type}{What type of residuals \code{"stan"}, \code{"stud"}}
}
\description{
Excludes values with high standardized / studentized residuals. If the largest residual > 2, values with residuals larger than 2, 2.5, 3, ... are excluded. If the largest residual < 2, values with 1:3 largest residuals are excluded (largest 3 standardized residuals is equivalent to the q-q plot definition of outliers in the regression diagnostics in the lm package)
}
