% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierExclusion.R
\name{.out.sdrule}
\alias{.out.sdrule}
\title{Standard deviation outlier definition}
\usage{
.out.sdrule(x, y)
}
\arguments{
\item{x}{Vector of values from which outliers should be excluded}

\item{y}{Vector of values from which outliers should be excluded}
}
\description{
Standard deviation outlier definition function: Takes a vector x, tries different standard deviation outlier rules (x > 2, 2.5, 3, ...) and returns all differing result vectors without the outliers.
}
