% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierExclusion.R
\name{.out.stemleaf}
\alias{.out.stemleaf}
\title{Stem and Leaf plot outlier definition}
\usage{
.out.stemleaf(x, y)
}
\arguments{
\item{x}{Vector of values from which outliers should be excluded}

\item{y}{Vector of values from which outliers should be excluded}
}
\description{
Stem and Leaf plot outlier definition function
}
