% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierExclusion.R
\name{.outHack}
\alias{.outHack}
\title{P-Hacking function for outlier exclusion in univariate linear regression}
\usage{
.outHack(df, x, y, which = c(1:12), strategy = "firstsig",
  alpha = 0.05, seed = 1234)
}
\arguments{
\item{df}{Data frame containing x and y variables as columns}

\item{x}{Location of x variable (predictor) in the data frame}

\item{y}{Location of y variable (criterion) in the data frame}

\item{which}{Which outlier definition methods? A numeric vector containing the chosen methods (1: boxplot, 2: stem&leaf, 3: standard deviation, 4: percentile, 5: studentized residuals, 6: standardized residuals, 7: DFBETA, 8: DFFITS, 9: Cook's D, 10: Mahalanobis distance, 11: Leverage values, 12: Covariance ratio)}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{alpha}{Significance level of the t-test (default: 0.05)}

\item{seed}{Initial seed for the random process}
}
\description{
Outputs a p-hacked p-value and a vector of all p-values that were computed in the process
}
