% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.report.group_lm}
\alias{.report.group_lm}
\title{Build one normalized reporting entry for a group effect in a linear model}
\usage{
.report.group_lm(formula, data, groupvar = "group", method = "ancova")
}
\arguments{
\item{formula}{Model formula}

\item{data}{Analysis data}

\item{groupvar}{Group variable name}

\item{method}{Analysis method}
}
\description{
Build one normalized reporting entry for a group effect in a linear model
}
