% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.report.multicat}
\alias{.report.multicat}
\title{Build one normalized reporting entry for a multi-category analysis}
\usage{
.report.multicat(group, y, method = "anova")
}
\arguments{
\item{group}{Grouping variable}

\item{y}{Criterion values}

\item{method}{Analysis method}
}
\description{
Build one normalized reporting entry for a multi-category analysis
}
