% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incorrectRounding.R
\name{.roundhack}
\alias{.roundhack}
\title{P-Hacking function for incorrect rounding}
\usage{
.roundhack(
  df,
  group,
  dv,
  roundinglevel,
  alternative = "two.sided",
  alpha = 0.05
)
}
\arguments{
\item{df}{Data frame}

\item{group}{Scalar defining location of the group vector in the data frame}

\item{dv}{Scalar defining location of dependent variable in the data frame}

\item{roundinglevel}{Highest p-value that is rounded down to 0.05}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater"). Here, \code{"greater"} tests whether the treatment or second group exceeds the control or first group.}

\item{alpha}{Significance level of the t-test (default: 0.05)}
}
\description{
Outputs a p-hacked p-value and the non-p-hacked-p-value
}
