% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.selectanalysis}
\alias{.selectanalysis}
\title{Select a p-hacked analysis from a vector of p-values}
\usage{
.selectanalysis(ps, strategy, alpha)
}
\arguments{
\item{ps}{Vector of p values}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{alpha}{Significance level (default: 0.05)}
}
\description{
Takes a vector of p-values and returns the index of the selected analysis.
}
