% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectEffects.R
\name{.selectEffects}
\alias{.selectEffects}
\title{P-Hacking function for selective reporting of effects}
\usage{
.selectEffects(df, dv, ivs, interactions = TRUE, ambitious = FALSE,
  alpha = 0.05)
}
\arguments{
\item{df}{Data frame containing dependent variable and composite score items as columns}

\item{dv}{Location of dependent variable in the data frame}

\item{ivs}{Location of independent variables in the data frame}

\item{interactions}{Should the model take interactions into account?}

\item{ambitious}{Ambitious p-hacking (smallest p value): TRUE/FALSE}

\item{alpha}{Significance level of the t-test (default: 0.05)}
}
\description{
P-Hacking function for selective reporting of effects
}
