% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.set_observed_fisherz}
\alias{.set_observed_fisherz}
\title{Transform a response to a requested observed Fisher-z association}
\usage{
.set_observed_fisherz(x, y, effect)
}
\arguments{
\item{x}{Predictor values}

\item{y}{Response values}

\item{effect}{Requested observed Fisher-z effect}
}
\description{
Transform a response to a requested observed Fisher-z association
}
