% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.set_observed_smd}
\alias{.set_observed_smd}
\title{Shift treatment values to a requested observed standardized mean difference}
\usage{
.set_observed_smd(control, treatment, effect)
}
\arguments{
\item{control}{Control group values}

\item{treatment}{Treatment group values}

\item{effect}{Requested observed standardized mean difference}
}
\description{
Shift treatment values to a requested observed standardized mean difference
}
