% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.sim.multcor}
\alias{.sim.multcor}
\title{Simulate multivariate correlated data for continuous variables}
\usage{
.sim.multcor(nobs, nvar, r, mu = 0, sd = 1, missing = 0)
}
\arguments{
\item{nobs}{Number of observations (rows) in the simulated data frame}

\item{nvar}{Number of variables (columns) in the data frame}

\item{r}{Desired correlation between the variables (integer)}

\item{mu}{Mean of the random data}

\item{sd}{Standard deviation of the random data}

\item{missing}{Proportion of missing values per variable (e.g., 0.2 = 20 percent)}
}
\description{
Outputs a data frame with correlated variables of defined length
}
