% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.sim.multregression}
\alias{.sim.multregression}
\title{Simulate one criterion and multiple predictors with shared study-level association}
\usage{
.sim.multregression(nobs, nvar, r, effect = 0, heterogeneity = 0, theta = NULL)
}
\arguments{
\item{nobs}{Number of observations}

\item{nvar}{Number of predictor variables}

\item{r}{Correlation between predictor variables}

\item{effect}{Mean Fisher-z effect size across studies}

\item{heterogeneity}{Between-study heterogeneity on the Fisher-z scale}

\item{theta}{Study-specific true effect size}
}
\description{
Simulate one criterion and multiple predictors with shared study-level association
}
