% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incorrectRounding.R
\name{.sim.ttest}
\alias{.sim.ttest}
\title{Simulate dataset for independent sample t-test}
\usage{
.sim.ttest(nobs.group)
}
\arguments{
\item{nobs.group}{Number of observations per group. Either a scalar or a vector with two elements.}
}
\description{
Outputs a data frame with two columns
}
