% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statAnalysis.R
\name{.statAnalysisHack}
\alias{.statAnalysisHack}
\title{P-Hacking function for exploiting different statistical analysis options}
\usage{
.statAnalysisHack(
  df,
  group,
  dv,
  strategy = "firstsig",
  alternative = "two.sided",
  alpha = 0.05
)
}
\arguments{
\item{df}{Data frame with one continuous independent variable and one continuous dependent variable}

\item{group}{Location of the grouping variable in the data frame}

\item{dv}{Location of the dependent variabl in the data frame}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater"). Here, \code{"greater"} tests whether the treatment or second group exceeds the control or first group.}

\item{alpha}{Significance level of the t-test}
}
\description{
P-Hacking function for exploiting different statistical analysis options
}
