% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableTransformation.R
\name{.varTransHack}
\alias{.varTransHack}
\title{P-Hacking function variable transformation in univariate linear regression}
\usage{
.varTransHack(df, x, y, transvar, strategy = "firstsig", alpha = 0.05)
}
\arguments{
\item{df}{Data frame containing x and y variables as columns}

\item{x}{Location of x variable (predictor) in the data frame}

\item{y}{Location of y variable (criterion) in the data frame}

\item{transvar}{Which variables should be transformed? Either "x" (for x variable), "y" (for y variable), or "xy" (for both)}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{alpha}{Significance level of the t-test (default: 0.05)}
}
\description{
Outputs a p-hacked p-value and a vector of all p-values that were computed in the process
}
