% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsShiny.R
\name{esplots}
\alias{esplots}
\title{Plot effect size distributions}
\usage{
esplots(
  simdat,
  EScolumn.hack,
  EScolumn.orig,
  titles = c(expression("Distribution of p-hacked effect sizes R"^2),
    expression("Distribution of original effect sizes R"^2))
)
}
\arguments{
\item{simdat}{Simulated data from one of the p-hacking simulation functions}

\item{EScolumn.hack}{Column number of hacked effect sizes}

\item{EScolumn.orig}{Column number of original effect sizes}

\item{titles}{Title of effect size plots}
}
\description{
Plot effect size distributions
}
