% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_elements_logit_single.R
\name{extract_elements_logit_single}
\alias{extract_elements_logit_single}
\title{extract_elements}
\usage{
extract_elements_logit_single(dv_models, x)
}
\arguments{
\item{x}{}
}
\value{
a single column tibble indicating significance and directionality of a logistic regression output
}
\description{
This is a backend helper function to interact with the phackr function.
}
