% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phackr.R
\name{phackr}
\alias{phackr}
\title{phackr}
\usage{
phackr(data, dvs, demos, logit = FALSE)
}
\arguments{
\item{data}{A survey data object. This can be created with the phackr_setup function}

\item{dvs}{A vector of dependent variables to run the same model across.}

\item{demos}{A vector of demographic variables and covariates to include in the ologit model.}

\item{logit}{A parameter to run logistic regressions. Defaults to ordinal logistic regression.}
}
\value{
Returns a tibble that can quickly be written to an excel sheet
}
\description{
The phackr function takes a dataset, a vector of dependent variables, and a vector of demographic variables and creates a tibble of multivariate significance from ordinal logit models.
}
\examples{
sheet1 <- phackr(data = survey_data, dvs = c("q2", "q3", "q4"), demos = c("race", "partyid", "gender"))

}
