% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phackr_setup.R
\name{phackr_setup}
\alias{phackr_setup}
\title{phackr_setup}
\usage{
phackr_setup(data, weight, caseid)
}
\arguments{
\item{data}{A dataframe, usually read from a dta file}

\item{weight}{The column name in the dataframe that would be the survey weight}

\item{caseid}{The column name in the dataframe that indicates caseid}
}
\value{
Returns a survey object
}
\description{
This function quickly sets up a survey object and transforms -99s, 77s, 98s, and 99s to NA.
}
\examples{
survey_data <- phackr_setup(data = "omnibus18", weight = "finalweight", caseid = "caseid")

}
