% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsShiny.R
\name{pplots}
\alias{pplots}
\title{Plot p-value distributions}
\usage{
pplots(simdat, alpha)
}
\arguments{
\item{simdat}{Simulated data from one of the p-hacking simulation functions}

\item{alpha}{Alpha level}
}
\description{
Plot p-value distributions
}
