% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favorableImputation.R
\name{sim.impHack}
\alias{sim.impHack}
\title{Simulate p-Hacking with different sorts of outlier definition missing value imputation}
\usage{
sim.impHack(
  nobs,
  missing,
  which = c(1:10),
  strategy = "firstsig",
  effect = 0,
  heterogeneity = 0,
  alpha = 0.05,
  iter = 1000,
  shinyEnv = FALSE
)
}
\arguments{
\item{nobs}{Integer giving number of observations}

\item{missing}{Percentage of missing values (e.g., 0.1 for 10 percent)}

\item{which}{Which imputation methods?  Either 5 random methods are chosen ("random") or a numeric vector containing the chosen methods (1: delete missing, 2: mean imputation, 3: median imputation, 4: mode imputation, 5: predictive mean matching, 6: weighted predictive mean matching, 7: sample from observed values, 8: Bayesian linear regression, 9: linear regression ignoring model error, 10: linear regression predicted values)}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{effect}{Mean effect size across studies on the Fisher-z scale}

\item{heterogeneity}{Between-study heterogeneity on the Fisher-z scale}

\item{alpha}{Significance level of the t-test (default: 0.05)}

\item{iter}{Number of simulation iterations}

\item{shinyEnv}{Is the function run in a Shiny session? TRUE/FALSE}
}
\description{
Outputs a data frame containing the p-hacked p-values (\code{ps.hack}), the original p-values (\code{ps.orig}), and a normalized reporting block from all iterations
}
