% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectiveReportingIV.R
\name{sim.multIVhack}
\alias{sim.multIVhack}
\title{Simulate p-Hacking with multiple independent variables}
\usage{
sim.multIVhack(
  nobs.group,
  nvar,
  r,
  regression = FALSE,
  strategy = "firstsig",
  effect = 0,
  heterogeneity = 0,
  iter = 1000,
  alternative = "two.sided",
  alpha = 0.05,
  shinyEnv = FALSE
)
}
\arguments{
\item{nobs.group}{Vector giving number of observations per group}

\item{nvar}{Number of independent variables (columns) in the data frame}

\item{r}{Desired correlation between the dependent variables (scalar)}

\item{regression}{Should the simulation be conducted for a regression analysis (TRUE) or a t-test? (FALSE)}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{effect}{Mean effect size across studies. For \code{regression = FALSE}, this is on the standardized mean-difference scale. For \code{regression = TRUE}, it is on the Fisher-z scale.}

\item{heterogeneity}{Between-study heterogeneity. For \code{regression = FALSE}, this is on the standardized mean-difference scale. For \code{regression = TRUE}, it is on the Fisher-z scale.}

\item{iter}{Number of simulation iterations}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater"). For \code{regression = FALSE}, \code{"greater"} tests whether the treatment or second group exceeds the control or first group. For \code{regression = TRUE}, it tests a positive association.}

\item{alpha}{Significance level of the t-test (default: 0.05)}

\item{shinyEnv}{Is the function run in a Shiny session? TRUE/FALSE}
}
\description{
Outputs a data frame containing the p-hacked p-values (\code{ps.hack}), the original p-values (\code{ps.orig}), and a normalized reporting block from all iterations
}
