% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incorrectRounding.R
\name{sim.roundhack}
\alias{sim.roundhack}
\title{Simulate p-hacking with incorrect rounding}
\usage{
sim.roundhack(
  roundinglevel,
  effect = 0,
  heterogeneity = 0,
  iter = 1000,
  alternative = "two.sided",
  alpha = 0.05,
  shinyEnv = FALSE
)
}
\arguments{
\item{roundinglevel}{Highest p-value that is rounded down to alpha}

\item{effect}{Mean effect size across studies}

\item{heterogeneity}{Between-study heterogeneity}

\item{iter}{Number of iterations}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater"). Here, \code{"greater"} tests whether the treatment or second group exceeds the control or first group.}

\item{alpha}{Significance level of the t-test (default: 0.05)}

\item{shinyEnv}{Is the function run in a Shiny session? TRUE/FALSE}
}
\description{
Outputs a data frame containing the p-hacked p-values (\code{ps.hack}), the original p-values (\code{ps.orig}), and a normalized reporting block from all iterations
}
