% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectEffects.R
\name{sim.selectEffects}
\alias{sim.selectEffects}
\title{Simulate p-hacking with selective reporting of effects
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations}
\usage{
sim.selectEffects(
  nobs,
  niv,
  riv,
  interactions = FALSE,
  strategy = "firstsig",
  alpha = 0.05,
  iter = 1000
)
}
\arguments{
\item{nobs}{Integer giving number of observations}

\item{niv}{Number of independent variables}

\item{riv}{Correlation between the independent variables}

\item{interactions}{Should the model take interactions into account?}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{alpha}{Significance level of the t-test (default: 0.05)}

\item{iter}{Number of simulation iterations}
}
\description{
Simulate p-hacking with selective reporting of effects
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations
}
