% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableTransformation.R
\name{sim.varTransHack}
\alias{sim.varTransHack}
\title{Simulate p-hacking with variable transformations}
\usage{
sim.varTransHack(
  nobs,
  transvar,
  testnorm = FALSE,
  strategy = "firstsig",
  effect = 0,
  heterogeneity = 0,
  alpha = 0.05,
  iter = 1000,
  shinyEnv = FALSE,
  empirical = FALSE
)
}
\arguments{
\item{nobs}{Integer giving number of observations}

\item{transvar}{Which variables should be transformed? Either "x" (for x variable), "y" (for y variable), or "xy" (for both)}

\item{testnorm}{Should variables only be transformed after a significant test for normality of residuals?}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{effect}{Mean effect size across studies on the Fisher-z scale}

\item{heterogeneity}{Between-study heterogeneity on the Fisher-z scale}

\item{alpha}{Significance level of the t-test (default: 0.05)}

\item{iter}{Number of simulation iterations}

\item{shinyEnv}{Is the function run in a Shiny session? TRUE/FALSE}

\item{empirical}{Should the observed initial effect be fixed to \code{effect}? If \code{TRUE}, \code{heterogeneity} is ignored.}
}
\description{
Outputs a data frame containing the p-hacked p-values (\code{ps.hack}), the original p-values (\code{ps.orig}), and a normalized reporting block from all iterations
}
