% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectiveReportingDV.R
\name{.multDVhack}
\alias{.multDVhack}
\title{P-Hacking function for multiple dependent variables}
\usage{
.multDVhack(df, dvs, group, strategy = "firstsig",
  alternative = "two.sided", alpha = 0.05)
}
\arguments{
\item{df}{Data frame with one group variable and multiple dependent variables}

\item{dvs}{Vector defining the DV columns (will be checked in given order)}

\item{group}{Scalar defining grouping column}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater")}

\item{alpha}{Significance level of the t-test}
}
\description{
Outputs a p-hacked p-value and a vector of all p-values that were computed in the process
}
