% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectiveReportingIV.R
\name{.multIVhack_reg}
\alias{.multIVhack_reg}
\title{P-Hacking function for multiple predictors in a regression}
\usage{
.multIVhack_reg(
  df,
  ivs,
  control,
  strategy = "firstsig",
  alternative = "two.sided",
  alpha = 0.05
)
}
\arguments{
\item{df}{Data frame containing a criterion variable and multiple predictor variables}

\item{ivs}{Location of the independent variables (predictors) in the data frame}

\item{control}{Location of the criterion in the data frame}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater")}

\item{alpha}{Significance level of the t-test (default: 0.05)}
}
\description{
Returns a p-hacked p-value and a vector of all p-values that were computed in the process
}
