% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionalStopping.R
\name{.optstop}
\alias{.optstop}
\title{Optional Stopping based on existing dataset}
\usage{
.optstop(df, group, y, n.min, n.max, step = 1,
  alternative = "two.sided", alpha = 0.05)
}
\arguments{
\item{df}{Data frame}

\item{group}{group Scalar defining grouping column}

\item{y}{Scalar defining location of dependent variable in the data frame}

\item{n.min}{Minimum sample size}

\item{n.max}{Maximum sample size}

\item{step}{Step size of the optional stopping (default is 1)}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater")}

\item{alpha}{Significance level of the t-test (default: 0.05)}
}
\description{
Outputs a p-hacked p-value and a non-p-hacked p-value based on the maximum sample size
}
