% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incorrectRounding.R
\name{.roundhack}
\alias{.roundhack}
\title{P-Hacking function for incorrect rounding}
\usage{
.roundhack(df, g1, g2, roundinglevel, alternative = "two.sided")
}
\arguments{
\item{df}{Data frame}

\item{g1}{Scalar defining location of first group in the data frame}

\item{g2}{Scalar defining location of second group in the data frame}

\item{roundinglevel}{Highest p-value that is rounded down to 0.05}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater")}
}
\description{
Outputs a p-hacked p-value and the non-p-hacked-p-value
}
