% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.sim.association}
\alias{.sim.association}
\title{Simulate two continuous variables with a study-level association}
\usage{
.sim.association(
  nobs,
  effect = 0,
  heterogeneity = 0,
  theta = NULL,
  missing = 0,
  empirical = FALSE
)
}
\arguments{
\item{nobs}{Number of observations}

\item{effect}{Mean Fisher-z effect size across studies}

\item{heterogeneity}{Between-study heterogeneity on the Fisher-z scale}

\item{theta}{Study-specific true effect size}

\item{missing}{Proportion of missing values per variable}

\item{empirical}{Should the observed initial effect be fixed to \code{effect}? If \code{TRUE}, \code{heterogeneity} is ignored.}
}
\description{
Simulate two continuous variables with a study-level association
}
