% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositeScores.R
\name{.sim.compscore}
\alias{.sim.compscore}
\title{Simulate data: Correlated composite score raw variables and one non-correlated dependent variable}
\usage{
.sim.compscore(nobs, ncompv, rcomp)
}
\arguments{
\item{nobs}{Integer giving number of observations}

\item{ncompv}{Integer giving number of variables to build the composite score}

\item{rcomp}{Integer giving correlation between the composite score variables}
}
\description{
Simulate data: Correlated composite score raw variables and one non-correlated dependent variable
}
