% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploitCovariates.R
\name{.sim.covariates}
\alias{.sim.covariates}
\title{Simulate data with (correlated) covariates}
\usage{
.sim.covariates(
  nobs.group,
  ncov,
  rcov,
  rcovdv,
  effect = 0,
  heterogeneity = 0,
  mu = 0,
  sd = 1,
  missing = 0,
  theta = NULL,
  empirical = FALSE
)
}
\arguments{
\item{nobs.group}{Vector with number of observations per group}

\item{ncov}{Number of continuous covariates in the simulated data frame}

\item{rcov}{Correlation between the covariates}

\item{rcovdv}{Correlation between covariates and dependent variable}

\item{effect}{Mean effect size across studies on the standardized mean-difference scale}

\item{heterogeneity}{Between-study heterogeneity on the standardized mean-difference scale}

\item{mu}{Mean of the random data}

\item{sd}{Standard deviation of the random data}

\item{missing}{Proportion of missing values per variable (e.g., 0.2 = 20 percent)}

\item{theta}{Study-specific true effect size}

\item{empirical}{Should the observed initial effect be fixed to \code{effect}? If \code{TRUE}, \code{heterogeneity} is ignored.}
}
\description{
Simulates a dependent variable that correlates with multiple (correlated) covariates as well as an independent IV
}
