% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploitCovariates.R
\name{.sim.covariates}
\alias{.sim.covariates}
\title{Simulate data with (correlated) covariates}
\usage{
.sim.covariates(nobs.group, ncov, rcov, rcovdv, mu = 0, sd = 1,
  missing = 0)
}
\arguments{
\item{nobs.group}{Vector with number of observations per group}

\item{ncov}{Number of continuous covariates in the simulated data frame}

\item{rcov}{Correlation between the covariates}

\item{rcovdv}{Correlation between covariates and dependent variable}

\item{mu}{Mean of the random data}

\item{sd}{Standard deviation of the random data}

\item{missing}{Proportion of missing values per variable (e.g., 0.2 = 20 percent)}
}
\description{
Simulates a dependent variable that correlates with multiple (correlated) covariates as well as an independent IV
}
