% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.sim.data}
\alias{.sim.data}
\title{Generic sampling function}
\usage{
.sim.data(
  nobs.group,
  effect = 0,
  heterogeneity = 0,
  theta = NULL,
  empirical = FALSE
)
}
\arguments{
\item{nobs.group}{Number of observations per group. Either a scalar or a vector with two elements.}

\item{effect}{Mean effect size across studies}

\item{heterogeneity}{Between-study heterogeneity}

\item{theta}{Study-specific true effect size}

\item{empirical}{Should the observed initial effect be fixed to \code{effect}? If \code{TRUE}, \code{heterogeneity} is ignored.}
}
\description{
Outputs a data frame with two columns
}
