% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.sim.data}
\alias{.sim.data}
\title{Generic sampling function}
\usage{
.sim.data(nobs.group)
}
\arguments{
\item{nobs.group}{Number of observations per group. Either a scalar or a vector with two elements.}
}
\description{
Outputs a data frame with two columns
}
