% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectiveReportingIV.R
\name{.sim.multIV}
\alias{.sim.multIV}
\title{Simulate dataset with multiple independent variables}
\usage{
.sim.multIV(nobs.group, nvar, r)
}
\arguments{
\item{nobs.group}{Scalar defining number of observations per group (if length = 2, the first element will be the size of the control group that stays the same)}

\item{nvar}{Number of independent variables in the data frame}

\item{r}{Desired correlation between the independent variables (scalar)}
}
\description{
Outputs data frame with multiple independent variables
}
