% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroupAnalysis.R
\name{.subgroupHack}
\alias{.subgroupHack}
\title{P-Hacking function for multiple subgroups analysis}
\usage{
.subgroupHack(
  df,
  iv,
  dv,
  subvars,
  alternative = "two.sided",
  strategy = "firstsig",
  alpha = 0.05
)
}
\arguments{
\item{df}{A matrix or data frame containing all relevant data}

\item{iv}{Integer specifying the location of the binary independent variable in the data frame}

\item{dv}{Integer specifying the location of the dependent variable in the data frame}

\item{subvars}{Vector specifying the location of the subgroup variables in the data frame}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater")}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{alpha}{Significance level of the t-test}
}
\description{
Outputs a p-hacked p-value and a vector of all p-values that were computed in the process
}
