% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositeScores.R
\name{sim.compscoreHack}
\alias{sim.compscoreHack}
\title{Simulate p-hacking with composite scores}
\usage{
sim.compscoreHack(
  nobs,
  ncompv,
  rcomp,
  ndelete,
  strategy = "firstsig",
  effect = 0,
  heterogeneity = 0,
  alpha = 0.05,
  iter = 1000,
  shinyEnv = FALSE,
  empirical = FALSE
)
}
\arguments{
\item{nobs}{Integer giving number of observations}

\item{ncompv}{Integer giving number of variables to build the composite score}

\item{rcomp}{Correlation between the composite score variables}

\item{ndelete}{How many items should be deleted from the scale at maximum?}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{effect}{Mean effect size across studies on the Fisher-z scale}

\item{heterogeneity}{Between-study heterogeneity on the Fisher-z scale}

\item{alpha}{Significance level of the t-test (default: 0.05)}

\item{iter}{Number of simulation iterations}

\item{shinyEnv}{Is the function run in a Shiny session? TRUE/FALSE}

\item{empirical}{Should the observed initial effect be fixed to \code{effect}? If \code{TRUE}, \code{heterogeneity} is ignored.}
}
\description{
Outputs a data frame containing the p-hacked p-values (\code{ps.hack}), the original p-values (\code{ps.orig}), and a normalized reporting block from all iterations
}
