% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositeScores.R
\name{sim.compscoreHack}
\alias{sim.compscoreHack}
\title{Simulate p-hacking with composite scores
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations}
\usage{
sim.compscoreHack(nobs, ncompv, rcomp, ndelete, ambitious = FALSE,
  alpha = 0.05, iter = 1000, seed = 1234)
}
\arguments{
\item{nobs}{Integer giving number of observations}

\item{ncompv}{Integer giving number of variables to build the composite score}

\item{rcomp}{Integer giving correlation between the composite score variables}

\item{ndelete}{How many items should be deleted from the scale at maximum?}

\item{ambitious}{Ambitious p-hacking (smallest p value): TRUE/FALSE}

\item{alpha}{Significance level of the t-test (default: 0.05)}

\item{iter}{Number of simulation iterations}

\item{seed}{Initial seed for the random process}
}
\description{
Simulate p-hacking with composite scores
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations
}
