% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploitCovariates.R
\name{sim.covhack}
\alias{sim.covhack}
\title{Simulate p-Hacking with multiple covariates}
\usage{
sim.covhack(
  nobs.group,
  ncov,
  rcov,
  rcovdv,
  interactions = FALSE,
  strategy = "firstsig",
  effect = 0,
  heterogeneity = 0,
  alpha = 0.05,
  iter = 1000,
  shinyEnv = FALSE
)
}
\arguments{
\item{nobs.group}{Vector with number of observations per group}

\item{ncov}{Number of continuous covariates in the simulated data frame}

\item{rcov}{Correlation between the covariates}

\item{rcovdv}{Correlation between covariates and dependent variable}

\item{interactions}{Should interaction terms be added to the ANCOVA models? TRUE/FALSE}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{effect}{Mean effect size across studies on the standardized mean-difference scale}

\item{heterogeneity}{Between-study heterogeneity on the standardized mean-difference scale}

\item{alpha}{Significance level of the t-test}

\item{iter}{Number of simulation iterations}

\item{shinyEnv}{Is the function run in a Shiny session? TRUE/FALSE}
}
\description{
Outputs a data frame containing the p-hacked p-values (\code{ps.hack}), the original p-values (\code{ps.orig}), and a normalized reporting block from all iterations
}
