% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploitCovariates.R
\name{sim.covhack}
\alias{sim.covhack}
\title{Simulate p-Hacking with multiple covariates
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations}
\usage{
sim.covhack(nobs.group, ncov, rcov, rcovdv, interactions = FALSE,
  ambitious = FALSE, alpha = 0.05, iter = 1000, seed = 1234)
}
\arguments{
\item{nobs.group}{Vector with number of observations per group}

\item{ncov}{Number of continuous covariates in the simulated data frame}

\item{rcov}{Correlation between the covariates}

\item{rcovdv}{Correlation between covariates and dependent variable}

\item{interactions}{Should interaction terms be added to the ANCOVA models? TRUE/FALSE}

\item{ambitious}{Ambitious p-hacking (smallest p value): TRUE/FALSE}

\item{alpha}{Significance level of the t-test}

\item{iter}{Number of simulation iterations}

\item{seed}{Initial seed for the random process}
}
\description{
Simulate p-Hacking with multiple covariates
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations
}
