% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploitCutoffs.R
\name{sim.cutoffHack}
\alias{sim.cutoffHack}
\title{Simulate p-Hacking for exploiting cutoff values
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations}
\usage{
sim.cutoffHack(nobs, ambitious = FALSE, alpha = 0.05, iter = 1000,
  seed = 1234)
}
\arguments{
\item{nobs}{Number of observations}

\item{ambitious}{Ambitious p-hacking (smallest p value): TRUE/FALSE}

\item{alpha}{Significance level of the t-test}

\item{iter}{Number of simulation iterations}

\item{seed}{Initial seed for the random process}
}
\description{
Simulate p-Hacking for exploiting cutoff values
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations
}
