% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploitCutoffs.R
\name{sim.cutoffHack}
\alias{sim.cutoffHack}
\title{Simulate p-Hacking for exploiting cutoff values
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations}
\usage{
sim.cutoffHack(nobs, strategy = "firstsig", alpha = 0.05, iter = 1000)
}
\arguments{
\item{nobs}{Number of observations}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{alpha}{Significance level of the t-test}

\item{iter}{Number of simulation iterations}
}
\description{
Simulate p-Hacking for exploiting cutoff values
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations
}
