% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectiveReportingDV.R
\name{sim.multDVhack}
\alias{sim.multDVhack}
\title{Simulate p-Hacking with multiple dependent variables}
\usage{
sim.multDVhack(
  nobs.group,
  nvar,
  r,
  strategy = "firstsig",
  effect = 0,
  heterogeneity = 0,
  iter = 1000,
  alternative = "two.sided",
  alpha = 0.05,
  shinyEnv = FALSE,
  empirical = FALSE
)
}
\arguments{
\item{nobs.group}{Vector giving number of observations per group}

\item{nvar}{Number of dependent variables (columns) in the data frame}

\item{r}{Desired correlation between the dependent variables (scalar)}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{effect}{Mean effect size across studies}

\item{heterogeneity}{Between-study heterogeneity}

\item{iter}{Number of simulation iterations}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater"). Here, \code{"greater"} tests whether the treatment or second group exceeds the control or first group.}

\item{alpha}{Significance level of the t-test (default: 0.05)}

\item{shinyEnv}{Is the function run in a Shiny session? TRUE/FALSE}

\item{empirical}{Should the observed initial effect be fixed to \code{effect}? If \code{TRUE}, \code{heterogeneity} is ignored.}
}
\description{
Outputs a data frame containing the p-hacked p-values (\code{ps.hack}), the original p-values (\code{ps.orig}), and a normalized reporting block from all iterations
}
