% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectiveReportingIV.R
\name{sim.multIVhack}
\alias{sim.multIVhack}
\title{Simulate p-Hacking with multiple independent variables}
\usage{
sim.multIVhack(
  nobs.group,
  nvar,
  r,
  regression = FALSE,
  strategy = "firstsig",
  iter = 1000,
  alternative = "two.sided",
  alpha = 0.05,
  shinyEnv = FALSE
)
}
\arguments{
\item{nobs.group}{Vector giving number of observations per group}

\item{nvar}{Number of independent variables (columns) in the data frame}

\item{r}{Desired correlation between the dependent variables (scalar)}

\item{regression}{Should the simulation be conducted for a regression analysis (TRUE) or a t-test? (FALSE)}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{iter}{Number of simulation iterations}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater")}

\item{alpha}{Significance level of the t-test (default: 0.05)}

\item{shinyEnv}{Is the function run in a Shiny session? TRUE/FALSE}
}
\description{
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations
}
