% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionalStopping.R
\name{sim.optstop}
\alias{sim.optstop}
\title{Simulate p-hacking with incorrect rounding}
\usage{
sim.optstop(
  n.min,
  n.max,
  step = 1,
  peek = NULL,
  alternative = "two.sided",
  iter = 1000,
  alpha = 0.05,
  shinyEnv = FALSE
)
}
\arguments{
\item{n.min}{Minimum sample size}

\item{n.max}{Maximum sample size}

\item{step}{Step size of the optional stopping (default is 1)}

\item{peek}{Determines how often one peeks at the data. Overrides step argument if not NULL.}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater")}

\item{iter}{Number of iterations}

\item{alpha}{Significance level of the t-test (default: 0.05)}

\item{shinyEnv}{Is the function run in a Shiny session? TRUE/FALSE}
}
\description{
Simulate p-hacking with incorrect rounding
}
