% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierExclusion.R
\name{sim.outHack}
\alias{sim.outHack}
\title{Simulate p-Hacking with different sorts of outlier definition}
\usage{
sim.outHack(
  nobs,
  which = c(1:12),
  strategy = "firstsig",
  alpha = 0.05,
  iter = 1000,
  shinyEnv = FALSE
)
}
\arguments{
\item{nobs}{Integer giving number of observations}

\item{which}{Which outlier detection methods?  Either 5 random methods are chosen ("random") or a numeric vector containing the chosen methods (1: boxplot, 2: stem&leaf, 3: standard deviation, 4: percentile, 5: studentized residuals, 6: standardized residuals, 7: DFBETA, 8: DFFITS, 9: Cook's D, 10: Mahalanobis distance, 11: Leverage values, 12: Covariance ratio)}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{alpha}{Significance level of the t-test (default: 0.05)}

\item{iter}{Number of simulation iterations}

\item{shinyEnv}{Is the function run in a Shiny session? TRUE/FALSE}
}
\description{
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations
}
