% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incorrectRounding.R
\name{sim.roundhack}
\alias{sim.roundhack}
\title{Simulate p-hacking with incorrect rounding}
\usage{
sim.roundhack(
  roundinglevel,
  iter = 1000,
  alternative = "two.sided",
  alpha = 0.05
)
}
\arguments{
\item{roundinglevel}{Highest p-value that is rounded down to 0.05}

\item{iter}{Number of iterations}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater")}

\item{alpha}{Significance level of the t-test (default: 0.05)}
}
\description{
Simulate p-hacking with incorrect rounding
}
