% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectEffects.R
\name{sim.selectEffects}
\alias{sim.selectEffects}
\title{Simulate p-hacking with selective reporting of effects
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations}
\usage{
sim.selectEffects(nobs, niv, riv, interactions = FALSE,
  ambitious = FALSE, alpha = 0.05, iter = 1000, seed = 1234)
}
\arguments{
\item{nobs}{Integer giving number of observations}

\item{niv}{Number of independent variables}

\item{riv}{Correlation between the independent variables}

\item{interactions}{Should the model take interactions into account?}

\item{ambitious}{Ambitious p-hacking (smallest p value): TRUE/FALSE}

\item{alpha}{Significance level of the t-test (default: 0.05)}

\item{iter}{Number of simulation iterations}

\item{seed}{Initial seed for the random process}
}
\description{
Simulate p-hacking with selective reporting of effects
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations
}
