% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statAnalysis.R
\name{sim.statAnalysisHack}
\alias{sim.statAnalysisHack}
\title{Simulate p-Hacking for exploiting different statistical analysis options}
\usage{
sim.statAnalysisHack(
  nobs.group,
  strategy = "firstsig",
  effect = 0,
  heterogeneity = 0,
  alternative = "two.sided",
  alpha = 0.05,
  iter = 1000,
  shinyEnv = FALSE
)
}
\arguments{
\item{nobs.group}{Number of observations per group. Either a scalar or a vector with 2 elements.}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{effect}{Mean effect size across studies}

\item{heterogeneity}{Between-study heterogeneity}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater"). Here, \code{"greater"} tests whether the treatment or second group exceeds the control or first group.}

\item{alpha}{Significance level of the t-test}

\item{iter}{Number of simulation iterations}

\item{shinyEnv}{Is the function run in a Shiny session? TRUE/FALSE}
}
\description{
Outputs a data frame containing the p-hacked p-values (\code{ps.hack}), the original p-values (\code{ps.orig}), and a normalized reporting block from all iterations
}
