% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statAnalysis.R
\name{sim.statAnalysisHack}
\alias{sim.statAnalysisHack}
\title{Simulate p-Hacking for exploiting different statistical analysis options}
\usage{
sim.statAnalysisHack(
  nobs.group,
  strategy = "firstsig",
  alternative = "two.sided",
  alpha = 0.05,
  iter = 1000
)
}
\arguments{
\item{nobs.group}{Number of observations per group. Either a scalar or a vector with 2 elements.}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater")}

\item{alpha}{Significance level of the t-test}

\item{iter}{Number of simulation iterations}
}
\description{
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations
}
