% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroupAnalysis.R
\name{sim.subgroupHack}
\alias{sim.subgroupHack}
\title{Simulate p-hacking with multiple subgroups}
\usage{
sim.subgroupHack(
  nobs.group,
  nsubvars,
  effect = 0,
  heterogeneity = 0,
  alternative = "two.sided",
  strategy = "firstsig",
  alpha = 0.05,
  iter = 1000,
  shinyEnv = FALSE,
  empirical = FALSE
)
}
\arguments{
\item{nobs.group}{Vector giving number of observations per group}

\item{nsubvars}{Integer specifying number of variables for potential subgroups}

\item{effect}{Mean effect size across studies}

\item{heterogeneity}{Between-study heterogeneity}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater"). Here, \code{"greater"} tests whether the treatment or second group exceeds the control or first group.}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{alpha}{Significance level of the t-test}

\item{iter}{Number of simulation iterations}

\item{shinyEnv}{Is the function run in a Shiny session? TRUE/FALSE}

\item{empirical}{Should the observed initial effect be fixed to \code{effect}? If \code{TRUE}, \code{heterogeneity} is ignored.}
}
\description{
Outputs a data frame containing the p-hacked p-values (\code{ps.hack}), the original p-values (\code{ps.orig}), and a normalized reporting block from all iterations
}
