% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroupAnalysis.R
\name{sim.subgroupHack}
\alias{sim.subgroupHack}
\title{Simulate p-hacking with multiple subgroups
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations}
\usage{
sim.subgroupHack(nobs.group, nsubvars, alternative = "two.sided",
  strategy = "firstsig", alpha = 0.05, iter = 1000, seed = 1234)
}
\arguments{
\item{nobs.group}{Vector giving number of observations per group}

\item{nsubvars}{Integer specifying number of variables for potential subgroups}

\item{alternative}{Direction of the t-test ("two.sided", "less", "greater")}

\item{strategy}{String value: One out of "firstsig", "smallest", "smallest.sig"}

\item{alpha}{Significance level of the t-test}

\item{iter}{Number of simulation iterations}

\item{seed}{Initial seed for the random process}
}
\description{
Simulate p-hacking with multiple subgroups
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations
}
