% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableTransformation.R
\name{sim.varTransHack}
\alias{sim.varTransHack}
\title{Simulate p-hacking with variable transformations
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations}
\usage{
sim.varTransHack(nobs, transvar, ambitious = FALSE, alpha = 0.05,
  iter = 1000, seed = 1234)
}
\arguments{
\item{nobs}{Integer giving number of observations}

\item{transvar}{Which variables should be transformed? Either "x" (for x variable), "y" (for y variable), or "xy" (for both)}

\item{ambitious}{Ambitious p-hacking (smallest p value): TRUE/FALSE}

\item{alpha}{Significance level of the t-test (default: 0.05)}

\item{iter}{Number of simulation iterations}

\item{seed}{Initial seed for the random process}
}
\description{
Simulate p-hacking with variable transformations
Outputs a matrix containing the p-hacked p-values (\code{ps.hack}) and the original p-values (\code{ps.orig}) from all iterations
}
